Merge changes from main repo

+++ src/mmapbitarray.h
@@ -63,13 +63,13 @@ MBArray * mbarray_Copy_Template(MBArray * src, char * filename, int perms);
 int mbarray_Update(MBArray * array, char * data, int size);
 /*MBArray * mbarray_Copy(MBarray * src, const char * filename);*/
 
-int mbarray_FileSize(MBArray * array);
+uint64_t mbarray_FileSize(MBArray * array);
 
 char * mbarray_CharData(MBArray * array);
 
 static inline size_t _vector_offset(MBArray * array, BTYPE bit)
 {
-    return array->preamblesize + bit / (sizeof(DTYPE) << 3);
+    return (size_t)(array->preamblesize + bit / (sizeof(DTYPE) << 3));
 }
 __attribute__((always_inline))
